mailAssistantByPrompts/requirements.txt
langchain_community
langchain
PyPDF
chainlit